factor out polynomial test helpers
