+++ cubic/shared/src/test/scala/org/hammerlab/math/polynomial/PolySolverTest.scala
@@ -1,28 +1,26 @@
 package org.hammerlab.math.polynomial
 
-import cats.syntax.show._
 import hammerlab.iterator._
 import org.hammerlab.Suite
-import org.hammerlab.math.format.showSuperscript
+import org.hammerlab.math.format.SigFigs
 import org.hammerlab.math.polynomial
-import spire.algebra.Ring
-import spire.implicits.{ DoubleAlgebra, IntAlgebra }
-import spire.syntax.all._
-import spire.math.Complex
+import spire.implicits._
 
-import math.abs
-import scala.Array.fill
+import scala.math.{ abs, exp }
+import scala.util.Random._
 
 abstract class PolySolverTest(N: Int)
   extends Suite {
 
   type D = Double
   type Real[T] = polynomial.Real[T]
-  type R[T] = polynomial.Result[T]
+  type R[T] = polynomial.Root[T]
 
   def M: Int
 
-  val casePrintInterval = 100
+  implicit var sigfigs: SigFigs = 3
+
+  val casePrintInterval = 1000
 
   val allImaginaryRootPairs =
     for {
@@ -30,40 +28,6 @@ abstract class PolySolverTest(N: Int)
       b ←  1 to M
     } yield
       ImaginaryRootPair[D](a, b)
-
-  def toComplex[T: Ring](r: R[T]): Seq[Complex[T]] =
-    r match {
-      case Real(r) ⇒ Seq(Complex(r))
-      case ImaginaryRootPair(a, b) ⇒ Seq(Complex(a, -b), Complex(a, b))
-    }
-
-  implicit class RootOps[T](r: R[T]) {
-    def complex(implicit e: Ring[T]): Seq[Complex[T]] = toComplex(r)
-  }
-
-  implicit class MultipleRootOps[T](r: (R[T], Int)) {
-    def complex(implicit e: Ring[T]): Seq[Complex[T]] = toComplex(r._1).flatMap(fill(r._2)(_))
-  }
-
-  import cats.Show
-  import Show.show
-
-  /**
-   * Pretty-print an imaginary-conjugate root-pair, possibly repeated as a multiple root (denoted by a superscript
-   */
-  implicit def showImagRoot[T]: Show[(ImaginaryRootPair[T], Int)] =
-    show {
-      case (pair, 1) ⇒ pair.toString
-      case (pair, degree) ⇒ s"($pair)${showSuperscript.show(degree)}"
-    }
-
-  implicit def showRealRoot[T]: Show[(Real[T], Int)] =
-    show {
-      case (real, 1) ⇒ real.toString
-      case (real, degree) ⇒ s"$real${showSuperscript.show(degree)}"
-    }
-
-
 
   /**
    * Iterate over the available imaginary-root conjugate-pairs (with integral real and imaginary parts in [-M, M], per
@@ -73,36 +37,6 @@ abstract class PolySolverTest(N: Int)
   def imaginaryRootPairs(num: Int) =
     allImaginaryRootPairs
       .unorderedSubsetsWithReplacement(num)
-
-  case class TestCase[T]( reals: Seq[(Real[T], Int)],
-                          imags: Seq[(ImaginaryRootPair[T], Int)],
-                          roots: Seq[Complex[T]],
-                         coeffs: Seq[T]) {
-    def scale = coeffs.head
-    override def toString: String =
-      s"roots: ${(reals.map(_.show) ++ imags.map(_.show)).mkString(", ")}\t coeffs: ${coeffs.mkString(" ")}"
-  }
-
-  object TestCase {
-    def apply[T: Ring](reals: Seq[(Real[T], Int)],
-                       imags: Seq[(ImaginaryRootPair[T], Int)],
-                       scale: T): TestCase[T] = {
-      val roots =
-        reals
-          .flatMap { case (ab, d) ⇒ fill(d)(ab) }
-          .flatMap(_.complex)
-        imags
-          .flatMap { case (ab, d) ⇒ fill(d)(ab) }
-          .flatMap(_.complex)
-
-      TestCase(
-        reals,
-        imags,
-        roots,
-        coeffs(roots, scale)
-      )
-    }
-  }
 
   /**
    * Iterator over all sets of `num` integer-roots in the range [-M,M] (including double/triple/quadruple roots)
@@ -132,55 +66,29 @@ abstract class PolySolverTest(N: Int)
 
         imags ← imaginaryRootPairs(numImaginaryRootPairs)
 
-        /** Expand multiple roots, convert to [[Complex]]s */
-        roots =
-          reals
-            .flatMap(_.complex) ++
-          imags
-            .flatMap { case (ab, d) ⇒ fill(d)(ab) }
-            .flatMap(_.complex)
-
         scale ← -scalesPerIteration to scalesPerIteration
         if scale != 0
 
       } yield
-        TestCase(
+        TestCase[D](
           reals,
           imags,
-          roots,
-          coeffs[D](roots, scale)
+          scale
         )
     )
 
-  import scala.util.Random._
-  import math.exp
-  def rnd(): Double = {
-    //exp(nextGaussian() * 5) * (if (nextBoolean()) 1 else -1)
-    nextGaussian()
-  }
+  def rnd(): Double = nextGaussian()
 
-  def coeffs[T: Ring](roots: Seq[Complex[T]], scale: T): Seq[T] = {
-    /** Infer the coefficients of the quartic equation from the roots */
-    scale ::
+  def logNormal(): Double =
+    exp(
+      nextGaussian() * 5
+    ) *
     (
-      for {
-        coeff ← 1 to N
-      } yield
-        roots
-          .unorderedSubsets(coeff)
-          .map(_.reduce(_ * _))
-          .reduce(_ + _) *
-            (
-              if (coeff % 2 == 0)
-                Ring[T].fromInt(1)
-              else
-                Ring[T].fromInt(-1)
-            )
+      if (nextBoolean())
+        1
+      else
+        -1
     )
-    .map(_.real)  // coefficients will always be real because all imaginary roots come in conjugate-pairs
-    .map(_ * scale)
-    .toList
-  }
 
   val iterationsPerRootShape = 100
   val scalesPerIteration = 3
@@ -206,8 +114,6 @@ abstract class PolySolverTest(N: Int)
         numRealRoots = N - numImagRoots
         realArities ← numRealRoots.unorderedPartitions
 
-        // _ = println(s"real/imaginary root-arities: ${realArities.mkString(",")} ${imagPairArities.mkString(",")}")
-
         // do 10 reps with each possible arity-distribution of [imaginary X real] roots
         _ ← 1 to iterationsPerRootShape
 
@@ -215,20 +121,12 @@ abstract class PolySolverTest(N: Int)
 
         imags = imagPairArities.map(ImaginaryRootPair(rnd(), abs(rnd())) → _)
 
-        roots =
-          reals
-            .flatMap(_.complex) ++
-          imags
-            .flatMap { case (pair, arity) ⇒ fill(arity)(pair) }
-            .flatMap(_.complex)
-
         scale ← (1 to scalesPerIteration).map(_ ⇒ rnd())
       } yield
         TestCase(
           reals,
           imags,
-          roots,
-          coeffs[D](roots, scale)
+          scale
         )
     )
   }
